Add files via upload
A/B testing is a data science technique used to compare marketing strategies, designs, or themes to achieve better results. In this project, we aim to enhance online user engagement for a bookstore by optimizing its website design. We will present users with two themes: Light Theme and Dark Theme. The focus will be on analyzing which theme boosts user engagement, influences Purchases, and improves the Conversion Rate—specifically, which theme leads to a higher number of user sign-ups. Through data-driven decisions, we aim to maximize the effectiveness of the bookstore's online presence.